feat(website): make warp landing the root, move legacy to /old
Promote the warp.dev-inspired landing (formerly /v2) to the root route
and relocate the previous landing to /old.

New root `/`:
- Light Acepe speed-glyph mark on dark with speckle texture (acepe-mark)
- Warp-style header with command chip, bento agent grid backed by static
  gradient art (no WebGL), static baked hero image (zero shaders)
- Copy: does not claim ACP is an open protocol

Legacy `/old`:
- Retains the single hero/demo shader; removes the per-feature
  FeatureCardShader instances and the bottom-CTA shader (static radial
  glow) to cut idle GPU cost
- Canonical URL updated to /old; perf contract test moved alongside

Also adds the "verify the DOM after UI changes" workflow learning.
